project_new/requirements.txt
torch
torchvision
flask
Pillow
